+++ lib/macros/macros.dart
@@ -39,6 +39,7 @@ export 'textDecorationStyle.dart';
 export 'textDirection.dart';
 export 'textInputAction.dart';
 export 'textStyleTheme.dart';
+export 'useQueryMacro.dart';
 export 'verticalDirection.dart';
 export 'wrapAlignment.dart';
 export 'wrapCrossAxis.dart';